+++ Assets/ML-Agents/Timers/step1_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":272.63091199999997,"total":272.7086835,"children":{"InitializeActuators":{"count":1,"self":0.0019962,"total":0.0019962,"children":null},"InitializeSensors":{"count":1,"self":0.00097749999999999985,"total":0.00097749999999999985,"children":null},"AgentSendState":{"count":587,"self":0.0069679,"total":0.0549222,"children":{"CollectObservations":{"count":118,"self":0.00099779999999999986,"total":0.00099779999999999986,"children":null},"WriteActionMask":{"count":118,"self":0,"total":0,"children":null},"RequestDecision":{"count":118,"self":0.0469565,"total":0.0469565,"children":null}}},"DecideAction":{"count":587,"self":0.012943099999999999,"total":0.012943099999999999,"children":null},"AgentAct":{"count":587,"self":0.0069448,"total":0.0069448,"children":null}},"gauges":{"Tank.CumulativeReward":{"count":2,"max":-1,"min":-1,"runningAverage":-1,"value":-1,"weightedAverage":-1}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1614304671","unity_version":"2019.4.18f1","command_line_arguments":"D:\\Unity\\2019.4.18f1\\Editor\\Unity.exe -projectpath D:\/Project\/UNITY\/ml-agents\/ML-tanks -useHub -hubIPC -cloudEnvironment production -hubSessionId de805e40-77b6-11eb-a214-b12b0d280286 -accessToken r_rTkH034sXJfS1Z0lW4x1DgHV-09phBj_NLddeYUuA00ef","communication_protocol_version":"1.3.0","com.unity.ml-agents_version":"1.7.2-preview","scene_name":"step1","end_time_seconds":"1614304944"}}
+{"count":1,"self":987.454976,"total":3962.9009237,"children":{"InitializeActuators":{"count":1,"self":0.0019944999999999997,"total":0.0019944999999999997,"children":null},"InitializeSensors":{"count":1,"self":0.0019952,"total":0.0019952,"children":null},"AgentSendState":{"count":1689134,"self":9.5241464,"total":37.214689199999995,"children":{"CollectObservations":{"count":337827,"self":1.9837272,"total":1.9837270999999999,"children":null},"WriteActionMask":{"count":337827,"self":0.8829578,"total":0.8829578,"children":null},"RequestDecision":{"count":337827,"self":4.8443784,"total":24.8238581,"children":{"AgentInfo.ToProto":{"count":337827,"self":2.2912564,"total":19.9794798,"children":{"GenerateSensorData":{"count":337827,"self":6.0744951999999994,"total":17.6882234,"children":{"RayPerceptionSensor.Perceive":{"count":337827,"self":11.613728,"total":11.613728199999999,"children":null}}}}}}}}},"DecideAction":{"count":1689134,"self":2921.6135168,"total":2921.6135528,"children":null},"AgentAct":{"count":1689134,"self":16.5041808,"total":16.504181499999998,"children":null},"AgentInfo.ToProto":{"count":1514,"self":0.0119394,"total":0.10855949999999999,"children":{"GenerateSensorData":{"count":1514,"self":0.033906,"total":0.0966201,"children":{"RayPerceptionSensor.Perceive":{"count":1514,"self":0.0627141,"total":0.0627141,"children":null}}}}}},"gauges":{"Tank.CumulativeReward":{"count":1514,"max":2,"min":-1,"runningAverage":0.542933166,"value":1,"weightedAverage":0.3398929}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1614321801","unity_version":"2019.4.18f1","command_line_arguments":"D:\\Unity\\2019.4.18f1\\Editor\\Unity.exe -projectpath D:\/Project\/UNITY\/ml-agents\/ML-tanks -useHub -hubIPC -cloudEnvironment production -hubSessionId de805e40-77b6-11eb-a214-b12b0d280286 -accessToken r_rTkH034sXJfS1Z0lW4x1DgHV-09phBj_NLddeYUuA00ef","communication_protocol_version":"1.3.0","com.unity.ml-agents_version":"1.7.2-preview","scene_name":"step1","end_time_seconds":"1614325764"}}
