+++ Assets/MLTank/TankObject/RootTank.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c62fd9a9b960f6f4886375ff261f7e6d
+guid: 6154673adffeccc4da2c8b36edbba29d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
