+++ Assets/MLTank/TankObject/tankA.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 085ad6073af07dc4b913062d6d3cbfb1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
