+++ Assets/MLTank/TankObject/tankA/tankA.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 5595901735023006244}
   - component: {fileID: 5595180185588595614}
   - component: {fileID: 8237024381198338392}
+  - component: {fileID: 7852541154004779302}
   m_Layer: 0
   m_Name: tankA
   m_TagString: Untagged
@@ -98,8 +99,11 @@ MonoBehaviour:
     maxStep: 0
   hasUpgradedFromAgentParameters: 1
   MaxStep: 0
+  target: {fileID: 0}
   tankTop: {fileID: 2335595501771705335}
   shotShell: {fileID: 6520900224311596044}
+  EnableMove: 0
+  launch_frequency_persec: 0.2
 --- !u!114 &8237024381198338392
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -114,6 +118,32 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   DecisionPeriod: 5
   TakeActionsBetweenDecisions: 1
+--- !u!114 &7852541154004779302
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 405897957522867300}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6bb6b867a41448888c1cd4f99643ad71, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_SensorName: RayPerceptionSensor
+  m_DetectableTags: []
+  m_RaysPerDirection: 3
+  m_MaxRayDegrees: 70
+  m_SphereCastRadius: 0.5
+  m_RayLength: 20
+  m_RayLayerMask:
+    serializedVersion: 2
+    m_Bits: 4294967291
+  m_ObservationStacks: 1
+  rayHitColor: {r: 1, g: 0, b: 0, a: 1}
+  rayMissColor: {r: 1, g: 1, b: 1, a: 1}
+  m_StartVerticalOffset: 0
+  m_EndVerticalOffset: 0
 --- !u!82 &560681495949756458
 AudioSource:
   m_ObjectHideFlags: 0
@@ -223,6 +253,7 @@ GameObject:
   - component: {fileID: 5665292812317323525}
   - component: {fileID: 6196317643914752884}
   - component: {fileID: 8198227294831911294}
+  - component: {fileID: 8737301756945008224}
   m_Layer: 0
   m_Name: top
   m_TagString: Untagged
@@ -317,6 +348,32 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b9192c3b6fe79754e9c96798e36a8f94, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &8737301756945008224
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2335595501771705335}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6bb6b867a41448888c1cd4f99643ad71, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_SensorName: RayPerceptionSensor
+  m_DetectableTags: []
+  m_RaysPerDirection: 3
+  m_MaxRayDegrees: 70
+  m_SphereCastRadius: 0.5
+  m_RayLength: 20
+  m_RayLayerMask:
+    serializedVersion: 2
+    m_Bits: 4294967291
+  m_ObservationStacks: 1
+  rayHitColor: {r: 1, g: 0, b: 0, a: 1}
+  rayMissColor: {r: 1, g: 1, b: 1, a: 1}
+  m_StartVerticalOffset: 0
+  m_EndVerticalOffset: 0
 --- !u!1 &7060799361397808285
 GameObject:
   m_ObjectHideFlags: 0
@@ -425,6 +482,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   latestPos: {x: 0, y: 0, z: 0}
+  EnableMove: 0
 --- !u!1 &8202926784524765729
 GameObject:
   m_ObjectHideFlags: 0
