+++ Assets/MLTank/TankObject/tankMe.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f4cc5a12d63dd64abc9dbd22f94216c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
