+++ Assets/MLTank/TankObject/tankMe/TankMe.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18df9bc77ba92424fae0d9bf8af7d7c2
+guid: 275d02b349714bb40b8c8fb3a03fc260
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
