+++ Assets/MLTank/TankObject/tankMe/TankMe.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4bb00efbebcc19549b998a2705076389
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
