+++ Assets/MLTank/TankObject/tankMe/test.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 186098ac62af9aa438153689ad621c5d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
