+++ Assets/Shell/ShotShell.cs
@@ -12,12 +12,13 @@ public class ShotShell : MonoBehaviour
     public void shotShell()
     {
         // 砲弾のプレハブを実体化（インスタンス化）する。
+        Debug.Log(this.transform.rotation);
         GameObject shell = Instantiate(shellPrefab, transform.position,this.transform.rotation );
         // 砲弾に付いているRigidbodyコンポーネントにアクセスする。
         Rigidbody shellRb = shell.GetComponent<Rigidbody>();
         // 前に飛ばす
         // Vector3 backward = new Vector3(0,0,-1);
-        shellRb.AddForce( transform.forward * shotSpeed);
+        shellRb.AddForce(transform.forward * shotSpeed);
         // 発射した砲弾を３秒後に破壊する。
         // （重要な考え方）不要になった砲弾はメモリー上から削除すること。
         //Destroy(shell, 3.0f);
@@ -30,7 +31,7 @@ public class ShotShell : MonoBehaviour
         // もしもSpaceキーを押したならば（条件）
         if ((Input.GetKeyDown(KeyCode.Space) || Input.GetMouseButtonDown(0)) && StageMaker.canMove )
         {
-            shotShell();
+            //shotShell();
         }
     }
 }
