+++ Assets/Shell/ShotShell.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 7295618439617545964}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 4}
-  m_LocalScale: {x: 2, y: 5, z: 2}
+  m_LocalScale: {x: 4, y: 25, z: 4}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -46,8 +46,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   shellPrefab: {fileID: 4915038115320194317, guid: 988574e4595729e42be7b4237e5df385,
     type: 3}
-  top: {fileID: 405897957522867300, guid: c138938fc5e554e46ac1a7e5794e2fdb, type: 3}
-  shotSpeed: 700
+  shotSpeed: 300
 --- !u!54 &7295618439617545955
 Rigidbody:
   m_ObjectHideFlags: 0
